Don't use specific cluster configurations here
Not required

For example: If there's cloudscale-rma-1 cluster, then
the API token should point to a project like
appcat-objects-cloudscale-rma-1. Also, since in the
resources (at least for S3) users can still choose a
different bucket region. Yet that bucket and
objectsuser should remain in the
appcat-objects-cloudscale-rma-1 project, even if the
bucket ends up in lpg.

For cross-provider-usage: Let's assume we have a
cluster exoscale-ch-gva-2 and we want to support
creating buckets on cloudscale's lpg region (e.g. for
offsite-backup buckets). This time we create a
cloudscale Project named
appcat-objects-exoscale-ch-gva-2 and it contains the
objects users and bucket of exoscale-ch-gva-2. There's
no link to cloudscale-lpg-2 in any way.