Fix nested defaulting for XRDs
Kubebuilder cannot set empty objects as a default for CRDs. We might be
able to fix this upstream, but I'll do this temporary hack to fix this
here.

This could be done with less duplication, but as we want to move away
from packages anyway, we should soon be able to do this in a cleaner way
in Jsonnet.
So, I don't think investing more time to make this prettier is helpful
at this point.