Fix connection secrets for VshnPostgreSQL
Setting the parameter `writeConnectionSecretToRef` on the claim should
result in a secret with the connection credentials named after the value
set in the `name` key of that parameter. However, the current
implementation writes the connection credentials to a hardcoded secret
called `<claim-name>-connection`. This leads to confusions for the end
user. This commit will fix this issue, by creating a temporary secret in
the namespace where the instance is placed and will then copy the
content to the secret specified in the claim.

Signed-off-by: Nicolas Bigler <nicolas.bigler@vshn.ch>