Adjust requests to the fair use ratio
The default requests have been adjusted. Also the requests for the
sidecars have been reduced a bit. For most of them the CPU has been
reduced from 250m to 100m and for the cluster-controller the memory
from 512Mi to 256Mi. As these are for standalone intances it should
be okay.