Add `writeConnectionSecretToRef` to Go types
When interacting with claims through Go we use the types we define in
the component.

This PR adds the `writeConnectionSecretToRef` field so we can interact
with the connection secret in Go. However this field is mananged by
Crossplane, wo strip out the field in Jsonnet before applying it to the
XRD.