Escape tags, so they support valid branch names
The `tag` field in the `appcat.image` object is used for two things:

1. It points to a git branch/tag
2. It's used for a docker image tag

The problem with this approach is that there are valid branch names that
aren't valid docker tags. This commit adds an escape function to build
the image that replaces invalid docker tag characters.

For example if `appcat.image.tag` is set to `add/test` it will use
`add/test` for the branch to render the manifests and the escaped string
`add_test` for the docker image tag.