Add billing support for APPPUiO Managed
This will add a recording rule which is slightly different between
appuio managed and appuio cloud. There's a new flag to handle this.

If the cluster is APPUiO Managed then the recording rule will get the
customer's ID hardcoded into the rule.

If the cluster is APPUiO Cloud, then the rule will use the provided
organization information.

Additionally, the query has been reworked so it's generic. It's now
able to query all current and future AppCat services as long as the
namespace where it's deployed contains a label of the pattern
`label_appuio_io_billing_name=~"appcat-.+"`.