Patch name of the inner ns object as well
Up to this point we only patched the name of the kube object, but forgot
to actually patch the name of the inner namespace object. This worked
because `provider-kubernetes` will automatically apply the name of the
outer object to the inner object during its reconcile. This got to be a
problem with 1.14 as the AppCat function would mess up the quotas if the
name was empty in the desired objects.

However this object does not adhere to our naming scheme of
`metadata.labels[crossplane.io/composite]+"-description"`. Unfortunately
we can't change that name now without re-creating the namespaces.