Explicitly specify readiness checks
On some clusters we have issues with the default readiness check, which
checks for the `Ready` condition on managed resources. However this
seems bugged on at least one of our clusters, as Crossplane claims the
check is invalid.

By explicitly adding it with an additional check, the check is accepted
again. Even though we add the exact same check again.