Remove the org label from the syn-appcat namespace
The organization label causes all pods in that namespace to have the
node selector set to a compute node (eg. flex).
However, we also have a node selector on the namespace level set to the
infra nodes.
This causes all pods to have two node selectors that are mutually
exclusive and results in unschedulable pods

Signed-off-by: Nicolas Bigler <nicolas.bigler@vshn.ch>