Implement proxy mode in component
Previously we had to to do a `make install-proxy` in the AppCat repo
to install the proxy configuration. This will not work very well with
the newly introduced ArgoCD in Kindev.

This change will make it easier to enable and push the proxy config, so
argo can deploy it. The component also checks for the `appcat_dev` fact
if it's not `true` it will refuse compilation.