Distinct name for the ArgoCD app
During an internal discussion we found out, that the default name for
the development ArgoCD app is the same as we use in production.

So if not careful it could happen that someone overrides the production
app.

With a distinct name we ensure that if someone is connected to a prod
cluster and executes `make push-golden` that we don't override the
production app but rather create an additional one. This additional one
will then simply error out as it can't connect to the local forgejo
instance.