Increase Crossplane poll rate
This change will decrease the load on the K8s API servers by 5-10x.

There's no impact on the time it takes until changes are reflected on
the claims or composites. Crossplane will reconcile unready composites
more aggressively. The 1h poll rate only applies to composites that
are considdered in a ready and synced state.