Increase Crossplane poll-interval to 1h
The Crossplane poll-interval is a bit different from the default
operator sync interval.

Crossplane manages so called managed-resources, which are opinionated
kubernetes objects. These manged resources are then in turn reconciled
by the providers. As they usually mange something in the cloud like
a bucket or dbaas, the state of the local managed resource might not
represent the actual state of the corresponding resource in the cloud.

That's why Crossplane triggers a reconcile on each of these managed
resource every `poll-interval`. This will result in quite a reconcile
cascade.

Now to reduce the load on the Kubernetes API we set the `poll-intervall`
to 1h. This only applies to composites that are in ready state. For
composites in unready state, the `poll-interval` is handled differently.
This will ensure that newly created or changed composites get their
changes reflected quicker.