Chose default backup region according to the OCP cluster zone
By default the component will now check what region the current OCP
cluster is on. Then it will compare against a list of regions, filter
out its own region and then take the first one.

It's also possible to override the region if the customer has a special
requirement for that.