Enable Compfunction Revisions integration
This commit enables the first iteration of comp function revisions.
It sets the composites to the `manual` `defaultCompositionUpdatePolicy`,
which prevents automatic roll out of new versions.

The maintenance has been enhanced to switch to the latest revisions
every time it runs. Making the whole rollout async and ensures that
restarts only happen during the maintenance.

Additionally there's a hotfix job, which loops over all claims and
composites to switch over the revision if there's an urgent reason to do
so.